+++ cli/cmd/migrate_test.go
@@ -8,21 +8,19 @@ import (
 )
 
 func TestMigrate(t *testing.T) {
-	t.Cleanup(func() {
-		os.RemoveAll("cloudquery.log")
-	})
-
 	_, filename, _, _ := runtime.Caller(0)
 	currentDir := path.Dir(filename)
 	testDataDir := path.Join(currentDir, "testdata")
 	cmd := NewCmdRoot()
-	cmd.SetArgs([]string{"migrate", testDataDir})
+	tmpDir := t.TempDir()
+	logFileName := path.Join(tmpDir, "cloudquery.log")
+	cmd.SetArgs([]string{"migrate", testDataDir, "--cq-dir", tmpDir, "--log-file-name", logFileName})
 	if err := cmd.Execute(); err != nil {
 		t.Fatal(err)
 	}
 
 	// check that log was written and contains some lines from the plugin
-	b, err := os.ReadFile("cloudquery.log")
+	b, err := os.ReadFile(logFileName)
 	if err != nil {
 		t.Fatal("failed to read cloudquery.log:", err)
 	}