+++ plugins/destination/test/client/client.go
@@ -2,6 +2,8 @@ package client
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"time"
 
 	"github.com/cloudquery/plugin-sdk/plugins"
@@ -14,11 +16,18 @@ import (
 type Client struct {
 	plugins.DefaultReverseTransformer
 	logger zerolog.Logger
+	spec   Spec
 }
 
 func New(ctx context.Context, logger zerolog.Logger, spec specs.Destination) (plugins.DestinationClient, error) {
+	var testConfig Spec
+	err := spec.UnmarshalSpec(&testConfig)
+	if err != nil {
+		return nil, fmt.Errorf("failed to unmarshal spec: %w", err)
+	}
 	return &Client{
 		logger: log.With().Str("module", "test").Logger(),
+		spec:   testConfig,
 	}, nil
 }
 
@@ -34,7 +43,14 @@ func (*Client) Migrate(ctx context.Context, tables schema.Tables) error {
 	return nil
 }
 
-func (*Client) Write(ctx context.Context, tables schema.Tables, res <-chan *plugins.ClientResource) error {
+//revive:disable We need to range over the channel to clear it, but revive thinks it can be removed
+func (c *Client) Write(ctx context.Context, tables schema.Tables, res <-chan *plugins.ClientResource) error {
+	if c.spec.ErrorOnWrite {
+		return errors.New("error_on_write is true")
+	}
+	for range res {
+		// do nothing
+	}
 	return nil
 }