+++ plugins/source/aws/codegen/services/clients.go
@@ -61,6 +61,8 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/neptune"
 	"github.com/aws/aws-sdk-go-v2/service/organizations"
 	"github.com/aws/aws-sdk-go-v2/service/qldb"
+	"github.com/aws/aws-sdk-go-v2/service/quicksight"
+	"github.com/aws/aws-sdk-go-v2/service/ram"
 	"github.com/aws/aws-sdk-go-v2/service/rds"
 	"github.com/aws/aws-sdk-go-v2/service/redshift"
 	"github.com/aws/aws-sdk-go-v2/service/resourcegroups"
@@ -148,6 +150,8 @@ var clients = []interface{}{
 	&neptune.Client{},
 	&organizations.Client{},
 	&qldb.Client{},
+	&quicksight.Client{},
+	&ram.Client{},
 	&rds.Client{},
 	&redshift.Client{},
 	&resourcegroups.Client{},