+++ plugins/source/aws/policies_v1/foundational_security/elbv2.sql
@@ -1,3 +1,0 @@
-\set check_id 'ELBv2.1'
-\echo "Executing check ELBv2.1"
-\ir ../queries/elb/elbv2_redirect_http_to_https.sql