+++ plugins/source/aws/resources/plugin/tables.go
@@ -62,6 +62,8 @@ import (
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/neptune"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/organizations"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/qldb"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/quicksight"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/ram"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/rds"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/redshift"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/resourcegroups"
@@ -315,7 +317,22 @@ func tables() []*schema.Table {
 		neptune.SubnetGroups(),
 		organizations.Accounts(),
 		qldb.Ledgers(),
+		quicksight.Analyses(),
+		quicksight.Dashboards(),
+		quicksight.DataSets(),
+		quicksight.DataSources(),
+		quicksight.Folders(),
+		quicksight.Groups(),
+		quicksight.Templates(),
+		quicksight.Users(),
+		ram.Principals(),
+		ram.Resources(),
+		ram.ResourceShares(),
+		ram.ResourceShareInvitations(),
+		ram.ResourceSharePermissions(),
+		ram.ResourceTypes(),
 		rds.Certificates(),
+		rds.EngineVersions(),
 		rds.ClusterParameterGroups(),
 		rds.ClusterSnapshots(),
 		rds.Clusters(),