+++ plugins/source/aws/resources/services/rds/engine_versions_fetch.go
@@ -1,0 +1,26 @@
+package rds
+
+import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/rds"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
+	"github.com/cloudquery/plugin-sdk/schema"
+)
+
+func fetchRdsEngineVersions(ctx context.Context, meta schema.ClientMeta, _ *schema.Resource, res chan<- interface{}) error {
+	c := meta.(*client.Client)
+	svc := c.Services().Rds
+
+	input := &rds.DescribeDBEngineVersionsInput{}
+
+	p := rds.NewDescribeDBEngineVersionsPaginator(svc, input)
+	for p.HasMorePages() {
+		response, err := p.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- response.DBEngineVersions
+	}
+	return nil
+}