+++ plugins/source/heroku/resources/services/add_on_webhook_delivery.go
@@ -13,9 +13,58 @@ import (
 func AddOnWebhookDeliveries() *schema.Table {
 	return &schema.Table{
 		Name:        "heroku_add_on_webhook_deliveries",
-		Description: `https://devcenter.heroku.com/articles/platform-api-reference#add-on-webhook-delivery-attributes`,
+		Description: `https://devcenter.heroku.com/articles/platform-api-reference#add-on-webhook-delivery`,
 		Resolver:    fetchAddOnWebhookDeliveries,
-		Columns:     []schema.Column{},
+		Columns: []schema.Column{
+			{
+				Name:     "created_at",
+				Type:     schema.TypeTimestamp,
+				Resolver: schema.PathResolver("CreatedAt"),
+			},
+			{
+				Name:     "event",
+				Type:     schema.TypeJSON,
+				Resolver: schema.PathResolver("Event"),
+			},
+			{
+				Name:     "id",
+				Type:     schema.TypeString,
+				Resolver: schema.PathResolver("ID"),
+				CreationOptions: schema.ColumnCreationOptions{
+					PrimaryKey: true,
+				},
+			},
+			{
+				Name:     "last_attempt",
+				Type:     schema.TypeJSON,
+				Resolver: schema.PathResolver("LastAttempt"),
+			},
+			{
+				Name:     "next_attempt_at",
+				Type:     schema.TypeTimestamp,
+				Resolver: schema.PathResolver("NextAttemptAt"),
+			},
+			{
+				Name:     "num_attempts",
+				Type:     schema.TypeInt,
+				Resolver: schema.PathResolver("NumAttempts"),
+			},
+			{
+				Name:     "status",
+				Type:     schema.TypeString,
+				Resolver: schema.PathResolver("Status"),
+			},
+			{
+				Name:     "updated_at",
+				Type:     schema.TypeTimestamp,
+				Resolver: schema.PathResolver("UpdatedAt"),
+			},
+			{
+				Name:     "webhook",
+				Type:     schema.TypeJSON,
+				Resolver: schema.PathResolver("Webhook"),
+			},
+		},
 	}
 }